shaders: allow soft-resetting pl_shader
Turns out this also solves the whole "but muh frequent re-allocations!"
problem, without needing some sort of ugly bstr pool. AKA just make the
user do it.

To avoid leaking memory for descriptor/variable names, we move those to
an explicit `tmp` object that we recycle when soft-resetting the shader.