vulkan: fix various issues
The signal mechanism wasn't powerful enough to handle multiple transfer
commands in succession. In theory, the new approach is now also flexible
enough to allow skipping events in the simple cases. Or perhaps, making
the use of events customizable as well.

Also fixes various parameter issues that arose because of the relaxed
dimensionality restrictions and generalized transfers etc.