colorspace: refactor matrix math etc.
The matrix math was separated from the color-space semantics and moved
to common.c; and all of the colorspace stuff was reworked.

In particular, the texture_mul business was fixed, and the decoding
matrix math simplified. (The more complete scenario which also does
texture_depth adjustments is now just a concatenation of the two
operations)