ra: generalize target_params even more
Turns out we can go all the way and just turn this into target_dummy,
allowing you to pass a real FBO to be used as the target dummy.

This actually helps mpv's use case, and others like it, because we
always create the FBO before the actual renderpass anyway, so we can
literally pass it and get free optimizations (in the case of vulkan).