shaders: implement HDR peak detection
Taken from mpv, with some modifications to fit the new system.

Not entirely sure I'm happy with the buffer building code. It might be a
good idea to add some sort of mid-level wrapper that makes this easier
for pl_shaders, but since the peak detection SSBO is the only current
user of this interface, doing it the ugly way is probably fine.