shaders: remove the dynamic variable renaming crap
This was brittle and ugly and slow and annoying anyway. Instead, turn
this into a user problem. Most users probably know what requirements
they have, no need to have this annoying code in their hot paths.

For the pl_dispatch system, we can just get this for free as long as you
`reset_frame` before starting over. This should be fine for the purposes
of pl_dispatch, which was never meant to be as flexible as pl_shader
anyway.