context: add more debugging-related checks
First of all, make sure we warn the user if NDEBUG was set; with a
suitably evil sounding env var guarding the presence of the warning.
Second, enable envvar-style leak reporting for talloc, much like mpv.

In both cases, I think using an env var is fine, since the code they
affect are not "thread safe" anyway, and they're not something the
library user really wants to care about in practice.