ra: add helpers for sorting and printing the formats
For sorting, we try a few heuristics but ultimately maintain the
guarantee that formats with "more" capabilities (popcount) are sorted
higher. This is probably what the user wants.

Printing the flags as a hex literal is sort of a cheap way out, but I
can't be bothered doing incremental buffer incrementation for the time
being. This is still better than the status quo.