meson: correctly install the headers
Using install_headers with subdirectories like that breaks completely.
Solve it by just using subdir, which also liberates us from having to
manually list all of the headers.

Due to a weird limitation of subdir(), we're forced to rename our
public include dir to `libplacebo`. So take this opportunity to also
move it to `include/libplacebo`. This has the pretty significant
advantage that you can just use -I to test directly from a local
libplacebo tree, which is a net gain overall.