context: improve the LC_NUMERIC locale check
Since e.g. `uselocale` can override `setlocale` for that thread, and we
have absolutely no way of querying this other than explicitly testing,
explicitly test.

In the case where the locale is specified wrong but it *seems* to work
regardless, just warn the user and hope they deal with it.