vulkan: correctly make buffer modifications host-visible
Even for coherent buffers, we technically need to ensure writes are made
visible to the host by including the HOST_READ_BIT in the dstAccessMask
of a buffer memory dependency. In order to avoid stalling the entire
pipeline, we do this "ahead of time" for any write performed on a buffer
that is marked as host-readable.