vulkan: slightly refactor context creation APIs
The default settings are changed (since the recent fixes have made async
compute stable to enable); and the public pl_vulkan object now also
exposes the generated device etc.

This is so users can end up combining it with VK_KHR_display or
something along those lines, as well as interop with other graphics
APIs, to wrap custom images. This also requires allowing the user to
enable custom extensions.

As an aside, fix a bug where the VK_KHR_surface extension name was not
correctly added when requesting surface compatibility.