shaders: implement dithering
Currently only supports ordered dither and white noise. I plan on adding
blue noise as well; perhaps based on the mpv fruit algorithm; but I
don't like the code for that so I'll probably reimplement blue noise
from scratch.

As a minor gotcha, the ordered dither code is fast and texture-free but
requires GLSL 130 or higher. Not sure what to do about this. I might set
the minimum GLSL version to 130 in general for libplacebo.