shaders: refactor randomness seeding
Instead of having the user pass in seed values, instead pass in a global
`index` on pl_shader generation, and use this to internally seed the
PRNG / perturb the dither matrix / etc.

This removes the highly state-dependent garbage from the *_params
structs, which is the right thing to do for the upcoming pl_renderer
options.