talloc: add refcounting API
This is a helper built on top of talloc, which maintains a refcounted
container. The function `ta_ref_attach` is a bit special, in that it
will attach a reference as a child of another talloc object such that
freeing the talloc object (or its children) will deref the reference.

Arguablly, it would probably be better to lump this into the extension
header and make it part of core TA, but I want to avoid touching TA as
much as possible.