shaders: add sh_subpass / shader merging
This adds the secondary shader as a subpass of the first shader;
exposing it as a function (much like pl_shader_finalize). There is
currently still no public shader merging API, but it could easily be
built on top of this.

This switches from void *tmp to ta_ref *tmp in order to avoid breaking
memory allocated by the attached objects.