osdep: explicitly set a category mask for newlocale
While on Linux it works fine to not specify one, as Linux will fill all
not given ones from the POSIX locale, on BSD using `xlocale.h` this does
not work.

When no base locale is given, BSD uses the current locale and fills not
matching bitmasks with values from that one, not with values from the
POSIX locale.