renderer: always align source planes to the refplane
Instead of the src_rect. This is needed if the src_rect is not an
integer crop. In theory, we could still use the src_rect here as long as
we make sure to take care of the correct src_rect in pass_scale_main,
but doing it this way is more consistent and less error-prone; although
slower.

Since this renderer has been WIP for way too long, I'll focus on the
easy solution rather than the ideal solution.