meson: update vulkan minimum version
This allows us to remove some work-arounds for old loader versions. I'm
not entirely sure what minimum version we actually require(*), but I
decided to go with 1.0.42 since that's what ships in e.g. ubuntu LTS and
is around a year old.

(*) We require at least 1.0.24 to solve the shader validation issue,
assuming the vulkan loader is built with in-tree SPIRV-Tools. But we
might require newer versions somewhere else, I don't have an easy way to
test this.