shaders: explicitly default the source colorimetry
In some cases, the destination colorspace is known but the source is
not. In these cases, the `needs_linear` etc. checks got unnecessarily
triggered, even though it was redundant. Even worse, due to floating
point error in the matrix math for producing what should have been an
identity matrix, we ended up accidentally triggering tone mapping!