filters: make the box() definition closer to the other kernels
This has to be radius 1 to work properly for orthogonal scaling, since
the code can't handle odd filter sizes properly. So just double the size
and adjust the kernel to compensate.

This is technically a waste of performance - it would be faster to fix
the code for separated scaling to cleanly support single-sample LUTs.
But that's very nontrivial, so this will have to do for now.