dispatch: correctly simulate the target rc for compute shaders
This is done by adjusting the output coordinate before writing to the
framebuffer. We also need to be extra careful not to write out of
bounds. To do that, just re-use the `out_scale` which we already have
for convenience.

It's worth noting that hard-coding the `ivec2 dir` like this could cause
a shader explosion (a variant for each possible combination), but since
it's limited to 2^2 = 4 it's probably not a big deal. I don't feel like
burning an ivec2 input just for this.