shaders: add pl_shader_encode_color
This is a functional counterpart to pl_shader_decode_color, used in the
output path where required. Since we want to avoid wasting a matrix on
this at all costs, aggressively check to see if the resulting matrix
would functionally be a no-op. (This will be the case most of the time)