renderer: encode to the target repr where needed
In addition, also use the bits structure's sample depth as authoritative
instead of the FBO format depth, since the caller might end up
discarding the extra bits.

Ideally, we would like to dither to the underlying color depth before
the bit shift, but this is difficult to arrange for, so ignore it for
now.