osdep: nuke printf.h
Instead of using shitty platform-specific hacks to get a
locale-invariant printf, we can just use our own printf routines for the
purposes of shader generation. This required making one slight
adjustment to one of the shaders in order to avoid the (now unsupported)
%%lu.

We now use the regular platform-provided snprintf for log messages and
other user-facing stuff, which should be fine. (We also technically use
it in a couple of places for snprintf with %d, but that should hopefully
not explode)