spirv: work around libshaderc locale dependence
Do this by explicitly clearing LC_NUMERIC before calling into
libshaderc. I'm not entirely sure whether libshaderc or glslang is to
blame here, but either way, as long as it works - I'm happy.

This is not that big of a deal performance-wise since shader compilation
is infrequent, and it also doesn't need messy global initialization
since we can just stuff the locale_t in with the struct priv.