renderer: only accept integer chroma scaling ratios
This works around the fact that subsampling oddly sized images results
in fractionally sized chroma planes that need to get rounded (typically
up).

Another way of phrasing it is that we only accept a regular chroma
siting cadence - since anything else would potentially lead itself to
nasty mosaicing artifacts as the chroma samples go in and out of phase
with the luma samples.