renderer: implement overlay rendering for pl_image as well
Closes #15

The most annoying thing about this is that I decided to try and
implement emulating the on-image OSD using the target OSD, which
involves some tricky manipulation. We could also just say "no" and
refuse to render on-image OSDs in such a situation, but it's not _that_
difficult either so this would probably be a strict improvement as far
as the status quo is concerned.

Anyway, I tested it a bit and it seems to work, so I'll just leave it
like this. If that code ever ends up causing problems in the future, we
can just delete it again.