renderer: completely refacter scaling
The bulk of the src_rect application is moved from the main scaling pass
to the pass_read_image step. This is needed in particular because we
can't retroactively apply the src_rect as part of the pass_read_image
pass when e.g. FBOs are unavailable, but also to allow us to make more
aggressive optimizations for skipping the main scaler on no-ops.

This is also much better for performance when cropping, since it avoids
all redundant work in the indirect pass, at the cost of possibly losing
some detail on edge pixels when cropping. (An acceptable trade-off IMO)

As a downside, we have to put in much more work to restore the use of
fractional src_rects, but this is a feature I care about.