ra: rename ra to pl_gpu, change ra_ to pl_
This is a very major rewrite operation, but all of the actual logic is
unaffected. The change is completely cosmetic.

The idea behind this is to avoid clashing the mpv ra_ namespace when
libplacebo eventually makes its way back into mpv, allowing it to
coexist with vo_gpu peacefully (at least for the transition period).

It's also sort of weirdly inconsistent with the rest of libplacebo
anyway.