vulkan: correctly enabled require device features
textureGatherOffset requires ImageGatherExtended, and the use of some
storage formats might require StorageImageExtendedFormats.

These need to be explicitly enabled at device creation time, so do so.
However, we don't want to pull in features that we don't use, so
explicitly whitelist the ones we want.

In case textureGatherOffset is unavailable, we explicitly clear the
gather offset to 0 (which means that feature is unsupported, by
definition).

As an aside, also correctly check to see if the gather offset is even
set, for sanity.