vulkan: implement texture format emulation
This required some pretty complicated logic, but I tested it and it
seems to be working. Basically, we can now just directly upload e.g.
rgb8 by treating it as rgba8. This is a gigantic hack, but it's much
faster than trying to CPU-convert it in cases where the data is already
readily available in packed form, e.g. in a mapped buffer.

Closes #9