shaders: avoid overflow in HDR peak detection
Instead of dividing out the number of workgroups per frame when
calculating the new signal peak, divide it out instead at the end of the
frame. This significantly lowers the size of the values we're
accumulating in `frames_sum` and hence `sum_total`, and prevents
overflow happening as a result of averaging together multiple frames
with high luminance values. Rename `sum` -> `avg` to better represent
the change.

As an aside, also clamp the detected peak to be at least the sdr_avg and
1.0 peak in order to avoid exploding fades-to-black.