shaders: enable support for HDR-in-SDR emulation
Since HDR TRC support in both hardware and software is flaky at best, we
can easily enable a sort of "soft HDR" by just calibrating the HDR
display to what's essentially a bright SDR curve with high contrast.

In this configuration, the new hdr_emulation feature allows treating
this as a true HDR display (with a normalized scale of 100 cd/m^2).