shaders: renderer: turn debanding into a real sampler
Debanding can now perform implicit bilinear scaling as well. We make use
of this to skip the indirection via FBO when there's either no support
for FBOs or advanced scalers were not requested.

We also use this to crop out only the area we're interested in, even
when re-scaling later.