renderer: factor out sampler type logic
Separate the decisions about what type of sampler to use from the logic
for actually applying them, so we can make queries about the sampler
type ahead of time (and use it to e.g. optionally take or not take code
paths).

This feels a lot cleaner than hackily reinventing this logic in multiple
places. Also, I have no idea why I ever wanted to pass in the `rx` and
`ry` instead of just calculating it from the rect. It makes no sense,
and it was probably wrong in a few places as well.