gpu: be more explicit about GPU limit types
Give them explicit size information and make them unsigned where it
matters. This solves an issue where platforms setting one or more of
these limits as UINT32_MAX would end up overflowing to (-1) due to the
representation as `int`.

Now the valid value range is a more explicit part of the API semantics,
with GPU implementations being required to convert and clamp as needed.

As an aside, also log the missing field max_buffer_texels.