vulkan: make sure texel buffers are large enough for emulation
Otherwise, emulated format upload results in an error (fortunately, the
libplacebo assert catches it).

This partially addresses #31, but it should be understood as an
immediate-term work-around. The longer-term solution is to use multiple
texel buffers.

It's worth noting that e.g. on my hardware, this prevents format
emulation from working, since the maximum 2D texture size is about twice
that of the maximum texel buffer size.