vulkan: add support for vulkan device extensions
In theory, we could also add instance extensions, but this makes it more
complicated when using an external VkInstance, which is a use case I
still care about. Since some of the VkDevice extensions we want to
optionally support depend on the VK_KHR_get_physical_device_properties2
instance extension, we need it to be available in order for the probing
to be reliable.

To make our lives easier, we can instead just require this extension,
since it's available almost everywhere anyway. If this list ends up
growing, we can make a better protocol here (a list of required and
optional extensions). But I don't particularly care for the time being.