vulkan: correctly align texture transfer buffer offsets
These need to be aligned to the texel size as well, rather than just
being a multiple of 4. Document and assert this fact for the user-facing
field.

We also need to make sure to align PL_BUF_TEX_TRANSFER buffers' memory
offsets ahead of time, so just use the LCM of all supported texture
formats in order to ensure we always get a valid allocation.

In addition to this, I also realized that PL_ALIGN cannot be used to
combine multiple non-PoT alignment requirements. So just define and use
`pl_lcm` (which we need anyway) to be on the safe side.