shaders/colorspace: make gamut warning work again
In 7aa14ac3 we made the tone mapping code hard-clip the output color
before passing it on down the line. But this sort of defeats the gamut
warning code, whose purpose was to detect clipped colors by checking for
anything >1.0 in the output.

Solve this dilemma by allowing a tiny bit of overshoot (1.01) and
checking for >1.005 instead. This shouldn't make a difference except a
tiny bit of discoloration when clipping, most likely not perceptible and
only happens with the clipping tone mapping modes anyway.