renderer: don't automatically use bicubic for built-in scaling
Using bicubic is actually a bad idea in the general case since it's not
only significantly slower compared to bilinear/nearest, but it also
forces an FBO pass for correct results.

Could maybe work around the latter thing, but in terms of the intent of
setting `NULL` here I think just forcing built-in scaling is the better
solution.