demos: add new extensive video filtering demo
This outlines multiple possible API styles from "simple and naive" all
the way to a complicated zero-copy asynchronous streaming API design.

Maybe in the future we could present a small benchmark to demonstrate
the improvements in throughput that are obtained from API #2? For now,
just put the code out there to serve as documentation.

Completely untested. It compiles, but that's about it. I'll probably
expand this sample to an actually working proof of concept state
soon-ish.