meson: exclude external libs on supported platforms
This prevents e.g. leaking glslang symbols, which is the main reason we
switched to the linker script in 6835ae5. Since that caused other
problems as well (notably failing on several platforms), and being just
as annoying to work with, I decided to revert that series of commits and
go back to the #pragma GCC approach - this at least works on all
platforms without any harm done. (If the pragma doesn't exist, it will
just get silently ignored)

And to solve the original motivation, we use the hide external libs flag.