vulkan: fix stall if buf_poll timeout is never > 0
The "optimization" in 223aa09 removed a "needless" flush that was very
much needed if the user simply never called pl_buf_poll with a nonzero
timeout. In this case, we never reap commands with `vk_poll_commands`,
and if the user does not `pl_gpu_flush` then we never even flush them to
begin with!

Fix both issues at the cost of having to flush every now and then.
Hopefully this will only matter for the first few frames anyway.