demos: add a benchmark/proof of concept to video-filtering.c
This doesn't really serve to do anything particularly useful, but it
demonstrates that the API can, in fact, be used - and more importantly,
it allows me to benchmark just how much faster API #2 is versus API #1
(and identify potential bottlenecks in order to get it faster).

This is also the first time I've ever benchmarked the actual
asynchronous texture upload/download process. Exciting stuff.