vulkan: fix render pass layout for newly created FBOs
This is actually a disturbingly common case, since most of the time the
pass will be created when the FBO is fresh as well. Fortunately, it only
matters when we need the layout for anything.

As an aside, maybe we should skip this 'optimization' step and just use
_UNDEFINED / _PREINITIALIZED depending on whether or not we need to
blend? Further testing warranted. But assuming it makes no difference,
the current code is fine. And assuming it makes a difference, the
current code would be better.