gpu: clarify and revise access rules on buffers
Due to the way the comment was formulated, it sounded like multiple
accesses by libplacebo to the same buffer were also forbidden while a
buffer was in use. However, this is simply not the case - we can do
texture up- and downloads all day, since these are synchronized
internally using semaphores.

So, tighten the scope of what's forbidden to explicitly writes performed
by the user (including pl_buf_read/write). Furthermore, while the
comment suggested that coalescing multiple writes was fine, this is
nonsensical - the result has to be undefined by definition since we
don't know the scope or ordering of the access. As a result, forbid it.