vulkan: perform buffer flushes *after* writing to the buffer
Right now, for some reason, when something involves visible writes, we
submit it as part of the buf_barrier. This is nonsensical, and only
works by coincidence. We need to actually include the write in the
source scope of the buf barrier. So we need to re-order the dependency
so that it happens after the buffer operation.

Accomplish this by splitting up the buffer barrier into two halves: one
synchronizing the current access w.r.t the previous access, and one
simply synchronizing the current access w.r.t host reads. (This does not
require changing buf_vk->current_access because no vulkan commands
correspond to host access, so there's nothing to transition out of)