vulkan: explicitly default PL_BUF_MEM_AUTO
The previous behavior assumed that the ordering in the vulkan driver's
memory types would be ordered by speed - but for any sort of buffer
involving host interaction, this is misleading. Sure, it might make the
actual buf->image copy faster, but it involves an additional memcpy from
host to device so the performance gain is negated.

To account for this, explicitly compensate for this by forcing
host-interactable buffer memory to be resident on the host.