colorspace: implement ITU-R BT.2100 ICtCp
This comes in two variants, PQ and HLG. Normally it should be easy
enough to infer this from the image's tagged TRC, but we don't really
have information about that in the shader stages concerning themselves
with the underlying image representation - so to be thorough, just
define both variants as separate enum constants. It should be easy
enough for users to infer the right type themselves.