gpu: move pl_buffer_var from gpu.h to shaders.h
This makes no real sense in gpu.h, since the pl_gpu doesn't actually
need this information in any way. It's only needed for the shader header
generation inside dispatch.h, so having it in shaders.h is the right
place.

Makes some code arguably a bit uglier, but whatever.