gpu: add pl_tex_export, replacing pl_vulkan_hold_external
This is similar to the old vulkan hold/release_external but more
generalized and more powerful. It also cleans up some of the semantics
related to external image interop. To avoid headaches, just delete the
old vulkan interop code in the same commit.

To make sure we don't keep around references to destroyed sync objects,
we refcount them. (This also means we don't need a lazy destructor,
since a sync object that's not in use can be destroyed immediately -
whereas a sync object that's in use will only be deref'd when the cmd
callback fires)