ta: assert successful list delink
ta_free ends up inadvertently modifying eh->children.next as part of
unlinking that value itself from teh linked list. But, clang doesn't
understand this - to clang, it looks as though we're freeing the same
value over and over again. Add a simple assertion to convince clang that
this is actually a different pointer.