3rdparty: use meson subprojects instead
Major refactor of the source code structure. Now uses meson subprojects
for "third party" dependencies (including bstr and talloc), and moves a
lot of the meson boilerplate from src/meson.build to the top-level
meson.build.

We choose the name 'xtalloc' to make sure we don't collide with samba
libtalloc, and to reflect the fact that our wrappers abort on OOM.