colorspace: improve the accuracy of LMS matrices
We were using both the old Von Kries LMS matrix (for color blindness
simulation), as well as the CIECAM97-derived non-linear Bradford matrix
for chromatic adaptation - despite not introducing the required
nonlinearity. (Note that ICC etc. also have the same error)

Switch to the revised CIECAM97 matrix which is designed with a linear
transform in mind. This also requires extending the logic of the color
blindness simulation matrices to ensure white continues mapping to white
- which previously only worked sort of by accident, and only to within
1e-3 accuracy.