vulkan: add support for loading instance-level functions
After postponing this issue previously because we had no permanent
context associated with user-provided instances, and no way of
retroactively knowing what instance-level extensions were provided, the
solution hit me on the head:

We can just load them as part of the device-level extensions that depend
on them. Duplicates in this list don't matter, since loading the same
pointer twice is idempotent.