malloc: determine the set of supported handles by trial
Just try asking the driver if it would be okay with a TRANSFER_DST
buffer for this handle type. If tha fails, so should certainly
everything else.

Unfortunately, we have no good way of testing which handle types are
supported for images. We could also use trials here, but we don't really
distinguish between images and buffers at all in the pl_gpu.handle_caps.

Hopefully this should normally either be supported for both buffers and
images, or for neither. It's probably best the user just relies on
`pl_tex_create` failing in environments where buffers are supported but
images are not.