meson: work-around meson bugs w.r.t subprojects
We can't link the subproejcts as dependencies directly, because meson
incorrectly emits the linker flags (and omits the objects). Solve it
manually instead by directly including the relevant object files.

This commit can and should be reverted as soon as the meson bugs are
fixed upstream, since it's an ugly hack.

cf. https://github.com/mesonbuild/meson/issues/4598