gpu: add import handle_caps
In prepartion for supporting importing external memory, we need to
define separate caps for import handle types. In theory, a Vulkan
implementation is free to only support a handle type for one
direction and not the other.

This also means we have a field for import caps for `pl_sync`. This is
always set to 0 for now as we don't have a use-case for importing
semaphores.