vulkan: add support for external memory with dma_buf fds
dma_buf fds are treated as a distinct type (vs OPAQUE_FD) in
Vulkan, and so we need to treat them separately too. While the
basic interactions are the same as for OPAQUE_FD, there is one
distinct difference, which is that dma_buf fds cannot be used
for external semaphores, so we have to separate the handle type
lists when probing for support.

Note that vkGetMemoryFdPropertiesKHR function import is currently
unused, but is necessary for importing dma_buf fds down the line.
While the method is part of the generic external_fd extension, the
spec says it cannot be used for OPAQUE_FD (seriously?) and so it's
really only relevant for dma_buf fds.