vulkan: add basic import test
This change adds a test that exports a texture and then imports it
back in again. This round-tripping is explicitly supported by the
Vulkan spec.

The test only runs for dma_buf handles because that's the only one
we have an import implementation for.

I also fixed the existing test which didn't properly check the
fd value of the buffer.