colorspace: add a sig_scale field to pl_color_space
This can be used to reinterpret what the brightness range of values
mean. A typical use case is for linear-light content which is
technically encoded as 0.0 - 1.0, but practically supposed to extend to
HDR values far beyond the typical SDR brightness. OpenEXR files in
integer mode are encoded like this, for example.

This also unifies/replaces the hacky old "hdr emulation" mode with the
new concept, it being simply applied to the target csp in this mode of
operation.