vulkan: implement proper buffer queue tracking
Turns out that this actually causes problems when e.g. trying to reuse
an SSBO between a compute shader and a fragment shader, which comes up
when using the new peak detection code.

One annoying thing about the implementation is that we will always need
to make sure that buf_flush happens after buf_signal, since the former
depends on the stage variable set by the latter. Maybe I'll change this
in the future if it gets annoying.

Fixes #49.