gpu: actually generalize `pl_tex_export`
We accidentally implemented this inside vulkan/gpu.c instead of the
general purpose wrapper code. That also meant we never tested the
`tex->params.export_handle` condition.

However, I realized that this condition is actually too restrictive for
our test framework, and working around it there would be sort of
annoying. So just drop the restriction.

I won't bother updating the API version for this change, since the
actual behavior hasn't changed. (And even if it had, it would only
matter for our own test framework)

As an aside, fix a bunch of related comments that still had outdated
field names in the documentation.