gpu: allow attaching arbitrary user data to tex/buf
Not implemented for the other types of objects, because the chances of a
user needing to associate unique data for objects is much less likely.
It helps for buf/tex in particular since the user can use this to e.g.
hold extra state that needs to be tracked for synchronization / external
API usage.

It's also motivated by the `pl_tex_dummy_create`, which can benefit from
allowing users to attach their own objects to dummy textures so they
know which one corresponds to what.